requirements/requirements-testing.txt
# PyTest for running the tests.
pytest>=2.8.0
pytest-django>=2.8.0
pytest-cov
pytest-sugar
coverage>=4.0
django-simple-captcha
django-wkhtmltopdf
django_coverage_plugin
django-test-plus
codecov
Pillow
django-bootstrap3
django-bootstrap-pagination
django-extensions
django-parsley
pygraphviz
django-extra-views
django-braces
django-wysiwyg-redactor
# django-ckeditor
-e git://github.com/django-ckeditor/django-ckeditor.git@master#egg=django-ckeditor

django-wysiwyg
django-model-utils
django-debug-toolbar
-e git://github.com/luzfcb/django-simple-history.git@wip-generic-views#egg=django-simple-history
-e git://github.com/maraujop/django-crispy-forms.git@dev#egg=django-crispy-forms

django-autocomplete-light
git+git://github.com/dcramer/django-devserver#egg=django-devserver
line_profiler